Рефакторит моки и вывод типов
