+++ src/components/app.tsx
@@ -6,22 +6,7 @@ import {Route, BrowserRouter, Routes, Link} from 'react-router-dom';
 import Favorites from '../pages/favorites';
 import Offer from '../pages/offer';
 
-
-type AppCard = {
-  id: number;
-  isPremium: boolean;
-  isFavorite: boolean;
-  imgUrl: string;
-  price: number;
-  tarification: string;
-  title: string;
-  type: string;
-  averageStars:number;
-}
-
-type AppCardProps = {
-  appData: AppCard[];
-}
+import type { Cards } from '../types';
 
 enum AppRoute {
   Root ='/',
@@ -31,11 +16,11 @@ enum AppRoute {
   EverythingElse ='*',
 }
 
-function App({appData}: AppCardProps) : JSX.Element {
+function App({cards}: Cards) : JSX.Element {
   return(
     <BrowserRouter>
       <Routes>
-        <Route path= {AppRoute.Root} element = { <MainPage cards = {appData} />} />
+        <Route path= {AppRoute.Root} element = { <MainPage cards = {cards} />} />
         <Route path= {AppRoute.LogIn} element= {<Login />}/>
         <Route path= {AppRoute.FavoriteItems} element= {<Favorites />} />
         <Route path= {AppRoute.OfferItem} element= {<Offer />}/>
