+++ src/types.ts
@@ -1,0 +1,17 @@
+type Card = {
+  id: number;
+  isPremium: boolean;
+  isFavorite: boolean;
+  imgUrl: string;
+  price: number;
+  tarification: string;
+  title: string;
+  type: string;
+  averageStars:number;
+}
+
+type Cards = {
+  cards: Card[];
+}
+
+export type {Card, Cards};
